IGetCityByIpAddress added

+++ WeatherInfoApi.ObjectModel/Responses/MainInfoResponse.cs
@@ -3,7 +3,7 @@
     public class MainInfoResponse
     {
         public double temp { get; set; }        
-        public int temp_min { get; set; }
+        public double temp_min { get; set; }
         public double temp_max { get; set; }
     }
 }
